Relocate the python interface helper include file to standard location
