+++ python/lsst/sphgeom/circle.cc
@@ -21,7 +21,7 @@
  */
 #include "pybind11/pybind11.h"
 
-#include "sphgeom.h"
+#include "lsst/sphgeom/python.h"
 
 #include "lsst/sphgeom/Box.h"
 #include "lsst/sphgeom/Circle.h"
