+++ python/lsst/sphgeom/lonLat.cc
@@ -21,7 +21,7 @@
  */
 #include "pybind11/pybind11.h"
 
-#include "sphgeom.h"
+#include "lsst/sphgeom/python.h"
 
 #include "lsst/sphgeom/LonLat.h"
 #include "lsst/sphgeom/Vector3d.h"
