+++ python/lsst/sphgeom/mq3cPixelization.cc
@@ -21,7 +21,7 @@
  */
 #include "pybind11/pybind11.h"
 
-#include "sphgeom.h"
+#include "lsst/sphgeom/python.h"
 
 #include "lsst/sphgeom/Mq3cPixelization.h"
