+++ python/lsst/sphgeom/pixelization.cc
@@ -21,7 +21,7 @@
  */
 #include "pybind11/pybind11.h"
 
-#include "sphgeom.h"
+#include "lsst/sphgeom/python.h"
 
 #include "lsst/sphgeom/Pixelization.h"
 #include "lsst/sphgeom/Region.h"
