+++ python/lsst/sphgeom/vector3d.cc
@@ -21,7 +21,7 @@
  */
 #include "pybind11/pybind11.h"
 
-#include "sphgeom.h"
+#include "lsst/sphgeom/python.h"
 
 #include "lsst/sphgeom/Angle.h"
 #include "lsst/sphgeom/UnitVector3d.h"
